refactor Pate.EventTrace to move general SymSequence functions to new What4.SymSequence module
